Geminabox with rsync to Google Cloud Storage
When the gem server starts last state is synced from the GCS bucket.
When the sync is completed the gem server and file system watcher is
started. The file system watcher will sync the gem data to GCS when the
file system changes.

To connect to the GCS bucket the following environment variables are
required:

- BUCKET (name of the bucket where the gem data will be saved)
- GCS_KEY (key of a service account for GCS base64 encoded)

To restrict access to the gem server authentication is required by
setting this required environment variables:

- USERNAME
- PASSWORD